Initial project setup with base and required files
This commit introduces the initial setup for the Kotlin library project. It includes the addition of the .gitignore file for ignoring unnecessary files and directories, the README.md and LICENSE.md files for providing the basic description and terms for the project and setting up the Kotlin coding standard in gradle.properties. A standard gradle wrapper is included for building the project. Furthermore, the JetBrains specific files (.idea/*) for project settings and configurations have also been added for ease of development. This setup is crucial for kick-starting the project development with all the necessary tools and configurations in place.