Add PairRequirementException.kt and corresponding test
Created `PairRequirementException.kt` custom class to handle pair requirement related exceptions. This class, which extends ConstraintException, offers a flexible way to generate error messages on-the-fly only when the exception is thrown thanks to a lazyMessage lambda.

Additionally, a unit test, `PairRequirementExceptionTest.kt`, was added to verify the correct functionality of `PairRequirementException.kt`, ensuring that an exception with a lazy message can be properly created.

The earlier definition of PairRequirementException in `Exceptions.kt` has been removed for better organization, as this exception now resides in its own file.