Revise constraints and enhance documentation in Ints package
Revamped constraints in the 'ints' package by revising the classes 'BeEqualTo', 'BeAtLeast', 'BeAtMost', and 'BePositive'. Made them data classes to aid comparability in tests. The method documentation now includes usage examples. Introduced 'BeEqualToTest' to check the behavior of 'BeEqualTo'. Also, the README was updated to provide a broad idea of using the library to validate user registration data. All these changes aim to improve the clarity and helpfulness of the constraints classes in the package, which would assist developers in better using the library.