Refactor constraint-related classes and tests
In this commit, multiple constraint-related classes are refactored. The DummyConstraint and ConstraintTest are removed as they were not needed anymore. The Constraint interface and some classes implementing it are changed. Now Constraint does not provide validate and validateNot methods and their implementation is moved to their usage in Jakt class. In addition, ConstraintException's child classes are renamed from RequirementException to ConstraintException for consistency. A number of test cases are updated to reflect these changes. These modifications are done to improve the constraints system and increase its usability and flexibility.