Remove Maven central publishing workflow
Deleted the GitHub Actions workflow 'release.yml' which was used for publishing to the Maven Central Repository. This change was necessary due to adjustments in our distribution strategy and we will no longer be using Maven central for publishing our package updates. Future updates will be published using a different method.