Add predicate support to HaveSize constraint, improve CompositeException, refactor Ranges and add CHANGELOG
Added predicate support to the HaveSize constraint to provide more versatile size checks. Also improved the CompositeException by ensuring that the list of exceptions is not empty. Simplified the Ranges utility class, removed deprecated functions and moved it to the appropriate package. Moreover, a CHANGELOG file was added following the industry standards to document future changes to the project.