Update dependencies and refactor tests in Jakt project
This commit updates several dependencies to their newer versions and modifies a test setup in the Jakt project. Specifically, the Kotest version for assertions, properties, and junit5 runner, along with the Dokka and Detekt plugins, have been updated. The usage of these dependencies in the respective config files has been adjusted accordingly. A JaktTest has also been refactored to opt into ExperimentalKotest and to modify the number of iterations within checkAll in a test case.